Move ffi, lib, and NULL to `gb.core`
