+++ README.md
@@ -146,14 +146,13 @@ Python-graphblas requires `numba` which enables compiling user-defined Python fu
 Example customized UnaryOp:
 ```python
 from graphblas import unary
-from graphblas.operator import UnaryOp
 
 def force_odd_func(x):
     if x % 2 == 0:
         return x + 1
     return x
 
-UnaryOp.register_new('force_odd', force_odd_func)
+unary.register_new('force_odd', force_odd_func)
 
 v = Vector.from_values([0, 1, 3], [1, 2, 3])
 w = v.apply(unary.force_odd).new()
