+++ graphblas/core/ss/vector.py
@@ -7,9 +7,10 @@ from suitesparse_graphblas.utils import claim_buffer, unclaim_buffer
 
 import graphblas as gb
 
-from ... import ffi, lib, monoid
+from ... import monoid
 from ...dtypes import _INDEX, INT64, UINT64, _string_to_dtype, lookup_dtype
 from ...exceptions import _error_code_lookup, check_status, check_status_carg
+from .. import NULL, ffi, lib
 from ..base import call
 from ..scalar import Scalar, _as_scalar
 from ..utils import _CArray, ints_to_numpy_buffer, libget, values_to_numpy_buffer, wrapdoc
@@ -20,7 +21,6 @@ from .prefix_scan import prefix_scan
 from .utils import get_order
 
 ffi_new = ffi.new
-NULL = ffi.NULL
 
 
 @njit
