+++ graphblas/tests/test_pygraphblas.py
@@ -17,7 +17,7 @@ def test_pygraphblas_matrix():  # pragma: no cover
     A = gb.Matrix.from_values([0, 1], [0, 1], [0, 1])
     pgA = A.to_pygraphblas()
     assert isinstance(pgA, pg.Matrix)
-    assert A.gb_obj is gb.ffi.NULL
+    assert A.gb_obj is gb.core.NULL
     pgA += 1
     A2 = gb.Matrix.from_pygraphblas(pgA)
     assert A2.isequal(gb.Matrix.from_values([0, 1], [0, 1], [1, 2]))
@@ -33,7 +33,7 @@ def test_pygraphblas_vector():  # pragma: no cover
     v = gb.Vector.from_values([0, 2], [0, 1])
     pgv = v.to_pygraphblas()
     assert isinstance(pgv, pg.Vector)
-    assert v.gb_obj is gb.ffi.NULL
+    assert v.gb_obj is gb.core.NULL
     pgv += 1
     v2 = gb.Vector.from_pygraphblas(pgv)
     assert v2.isequal(gb.Vector.from_values([0, 2], [1, 2]))
